ORTGuardImpl: finish implementing this
Unfortunately due to how autograd works, the number of devices that
the guard claims to support is the number of potential threads that
autograd will attempt to light up to perform work against a device key,
so we cannot really use a device index as a private metadata space
for ORT EP orchestration.

So, back out this concept and simply claim support for a single ORT
device. We will need to figure out how to route to the correct EP
via another (out of band) mechanism, probably some API that binds
an EP to the process itself and not a thread, since PyTorch may
run autograd code on any number of background threads.