Generator: generate unique function names for registration overloads
We can't register native op functions with the same name since the
compiler will not have enough information at the registration site to
pick the correct overload. For example, the C++ functions for fmod
will be named aten::fmod__Tensor and aten::fmod__Scalar now.

Also generate C++ namespaces for ops, so our ATen ops for example
will be in torch_ort::eager::aten. This will allow for custom ops
and non-ATen ops to be in proper namespaces going forward (and not
assumed to be ATen).